refactor(web): extract admin feedback section into a child component
Behaviour-preserving split of admin-page.component.ts (771 → 172 LOC) to meet
the ≤250 LOC rule:
- extract the self-contained feedback table (state, sort, load + row actions:
  mark read, delete, create GitHub issue) into AdminFeedbackSectionComponent
- externalize both components' templates/styles to .html/.scss
- move the feedback unit tests to the child spec; parent spec keeps the
  user-table/dialog tests

i18n message IDs unchanged. AdminFeedbackSectionComponent is 179 LOC.

Refs #456
