+++ web/src/app/admin/admin-feedback-section.component.scss
@@ -1,0 +1,44 @@
+h2 {
+  margin-top: 48px;
+  margin-bottom: 16px;
+}
+.filter-row {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+  margin-bottom: 16px;
+}
+.spinner-container {
+  display: flex;
+  justify-content: center;
+  padding: 48px;
+}
+.error-text {
+  color: var(--mat-sys-error);
+}
+.uid-chip {
+  font-family: monospace;
+  font-size: 0.85em;
+  background: var(--mat-sys-surface-variant);
+  padding: 2px 6px;
+  border-radius: 4px;
+}
+mat-table {
+  width: 100%;
+}
+.message-cell {
+  white-space: pre-wrap;
+  word-break: break-word;
+  max-width: 400px;
+}
+.anon-icon {
+  opacity: 0.5;
+}
+.feedback-read {
+  opacity: 0.55;
+}
+.feedback-actions-cell {
+  display: flex;
+  gap: 4px;
+  justify-content: flex-end;
+}
