+++ web/src/app/admin/admin-page.component.scss
@@ -1,0 +1,60 @@
+.admin-page {
+  padding: 24px;
+  max-width: 1200px;
+  margin: 0 auto;
+}
+h1 {
+  margin-bottom: 24px;
+}
+.bulk-card {
+  margin-bottom: 24px;
+}
+.bulk-card mat-form-field {
+  margin-top: 16px;
+}
+.bulk-hint {
+  color: var(--mat-sys-on-surface-variant);
+  font-size: 0.85em;
+  margin-top: 8px;
+}
+.bulk-result {
+  font-weight: 500;
+}
+.filter-row {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+  margin-bottom: 16px;
+}
+.spinner-container {
+  display: flex;
+  justify-content: center;
+  padding: 48px;
+}
+.error-text {
+  color: var(--mat-sys-error);
+}
+.uid-chip {
+  font-family: monospace;
+  font-size: 0.85em;
+  background: var(--mat-sys-surface-variant);
+  padding: 2px 6px;
+  border-radius: 4px;
+}
+.anon-icon-warn {
+  color: var(--mat-sys-error);
+}
+mat-table {
+  width: 100%;
+}
+.clickable-row {
+  cursor: pointer;
+}
+.clickable-row:hover {
+  background: var(--mat-sys-surface-variant);
+}
+.clickable-row:focus-visible {
+  outline: 2px solid var(--mat-sys-primary);
+  outline-offset: -2px;
+  background: var(--mat-sys-surface-variant);
+}
