+++ web/src/app/admin/admin-page.component.spec.ts
@@ -9,7 +9,6 @@ import {
 import { of } from 'rxjs';
 import { AdminPageComponent } from './admin-page.component';
 import { AdminUser } from './admin-page.models';
-import { DeleteFeedbackDialogComponent } from './delete-feedback-dialog.component';
 import { DeleteUserDialogComponent } from './delete-user-dialog.component';
 import { UserDetailsDialogComponent } from './user-details-dialog.component';
 
@@ -46,38 +45,26 @@ describe('AdminPageComponent', () => {
     return { afterClosed: () => of(result) } as MatDialogRef<unknown, T>;
   }
 
-  const baseFeedback = [
-    {
-      id: 'fb-1',
-      name: 'Alice',
-      email: 'alice@example.com',
-      message: 'Nice app!',
-      userId: 'user-1',
-      createdAt: '2026-04-09T10:00:00.000Z',
-      userAgent: null,
-      read: false,
-      githubIssueUrl: null,
-    },
-    {
-      id: 'fb-2',
-      name: null,
-      email: null,
-      message: 'Bug report: button does nothing',
-      userId: null,
-      createdAt: '2026-04-08T09:00:00.000Z',
-      userAgent: null,
-      read: false,
-      githubIssueUrl: null,
-    },
-  ];
+  const sampleUser: AdminUser = {
+    uid: 'user-1',
+    displayName: 'Alice',
+    email: 'alice@example.com',
+    anonymous: false,
+    pushupCount: 12,
+    lastEntry: '2026-04-09T10:00:00.000Z',
+    createdAt: '2026-01-01T00:00:00.000Z',
+    role: null,
+  };
 
+  // The embedded <app-admin-feedback-section> loads feedback on init, so every
+  // render needs the adminListFeedback callable stubbed alongside the users one.
   async function createComponent(
-    feedback = baseFeedback,
+    users: AdminUser[] = [],
     extraCallables: CallableRecord[] = []
   ): Promise<void> {
     setupCallables([
-      { name: 'adminListUsers', impl: async () => ({ data: [] }) },
-      { name: 'adminListFeedback', impl: async () => ({ data: feedback }) },
+      { name: 'adminListUsers', impl: async () => ({ data: users }) },
+      { name: 'adminListFeedback', impl: async () => ({ data: [] }) },
       ...extraCallables,
     ]);
 
@@ -103,213 +90,82 @@ describe('AdminPageComponent', () => {
     vi.clearAllMocks();
   });
 
-  describe('feedback list initialization', () => {
-    it('loads feedback on init via adminListFeedback callable', async () => {
-      await createComponent();
-      expect(component.feedbackList().length).toBe(2);
-      expect(component.feedbackList()[0].id).toBe('fb-1');
+  describe('user list initialization', () => {
+    it('should load users on init via the adminListUsers callable', async () => {
+      // given / when
+      await createComponent([sampleUser]);
+
+      // then
+      expect(component.users().length).toBe(1);
       expect(httpsCallable).toHaveBeenCalledWith(
         expect.anything(),
-        'adminListFeedback'
+        'adminListUsers'
       );
-    });
-
-    it('keeps the list empty when the Cloud Function returns empty data', async () => {
-      await createComponent([]);
-      expect(component.feedbackList()).toEqual([]);
-    });
-  });
-
-  describe('deleteFeedback', () => {
-    it('opens the confirmation dialog before deleting', async () => {
-      await createComponent();
-      dialogOpenSpy.mockReturnValue(stubDialogRef(false));
-
-      await component.deleteFeedback(baseFeedback[0]);
-
-      expect(dialogOpenSpy).toHaveBeenCalledWith(
-        DeleteFeedbackDialogComponent,
-        expect.objectContaining({
-          data: { name: 'Alice', message: 'Nice app!' },
-        })
-      );
-    });
-
-    it('does NOT call adminDeleteFeedback when the user cancels', async () => {
-      const deleteCallable = vi.fn(async () => ({ data: { ok: true } }));
-      await createComponent(baseFeedback, [
-        { name: 'adminDeleteFeedback', impl: deleteCallable },
-      ]);
-      dialogOpenSpy.mockReturnValue(stubDialogRef(false));
-
-      await component.deleteFeedback(baseFeedback[0]);
-
-      expect(deleteCallable).not.toHaveBeenCalled();
-      expect(component.feedbackList().length).toBe(2);
-    });
-
-    it('calls adminDeleteFeedback and removes the row when the user confirms', async () => {
-      const deleteCallable = vi.fn(async () => ({ data: { ok: true } }));
-      await createComponent(baseFeedback, [
-        { name: 'adminDeleteFeedback', impl: deleteCallable },
-      ]);
-      dialogOpenSpy.mockReturnValue(stubDialogRef(true));
-
-      await component.deleteFeedback(baseFeedback[0]);
-
-      expect(deleteCallable).toHaveBeenCalledWith({ feedbackId: 'fb-1' });
-      expect(component.feedbackList().map((f) => f.id)).toEqual(['fb-2']);
-      expect(component.feedbackActionError()).toBeNull();
-    });
-
-    it('shows an error message and keeps the row when the Cloud Function fails', async () => {
-      const deleteCallable = vi.fn(async () => {
-        throw new Error('permission-denied');
-      });
-      await createComponent(baseFeedback, [
-        { name: 'adminDeleteFeedback', impl: deleteCallable },
-      ]);
-      dialogOpenSpy.mockReturnValue(stubDialogRef(true));
-
-      await component.deleteFeedback(baseFeedback[0]);
-
-      expect(component.feedbackActionError()).toBe('permission-denied');
-      expect(component.feedbackList().length).toBe(2);
-      expect(component.isFeedbackActionLoading('fb-1')).toBe(false);
-    });
-
-    it('resets the per-row loading state after a successful delete', async () => {
-      const deleteCallable = vi.fn(async () => ({ data: { ok: true } }));
-      await createComponent(baseFeedback, [
-        { name: 'adminDeleteFeedback', impl: deleteCallable },
-      ]);
-      dialogOpenSpy.mockReturnValue(stubDialogRef(true));
-
-      await component.deleteFeedback(baseFeedback[0]);
-
-      expect(component.isFeedbackActionLoading('fb-1')).toBe(false);
-    });
-  });
-
-  describe('markFeedbackRead', () => {
-    it('flips the read flag for the matching row on success', async () => {
-      const markCallable = vi.fn(async () => ({ data: { ok: true } }));
-      await createComponent(baseFeedback, [
-        { name: 'adminMarkFeedbackRead', impl: markCallable },
-      ]);
-
-      await component.markFeedbackRead(baseFeedback[0], true);
-
-      expect(markCallable).toHaveBeenCalledWith({
-        feedbackId: 'fb-1',
-        read: true,
-      });
-      expect(component.feedbackList()[0].read).toBe(true);
     });
   });
 
   describe('openDetailsDialog', () => {
-    const sampleUser: AdminUser = {
-      uid: 'user-1',
-      displayName: 'Alice',
-      email: 'alice@example.com',
-      anonymous: false,
-      pushupCount: 12,
-      lastEntry: '2026-04-09T10:00:00.000Z',
-      createdAt: '2026-01-01T00:00:00.000Z',
-      role: null,
-    };
-
-    it('opens UserDetailsDialogComponent with the clicked user as data', async () => {
+    it('should open UserDetailsDialogComponent with the clicked user as data', async () => {
+      // given
       await createComponent();
       dialogOpenSpy.mockReturnValue(stubDialogRef(undefined));
 
+      // when
       component.openDetailsDialog(sampleUser);
 
+      // then
       expect(dialogOpenSpy).toHaveBeenCalledWith(
         UserDetailsDialogComponent,
         expect.objectContaining({ data: sampleUser })
       );
     });
 
-    it('opens the details dialog when a non-actions cell of a rendered row is clicked', async () => {
-      // Render a row so the template (click) wiring is exercised.
-      setupCallables([
-        { name: 'adminListUsers', impl: async () => ({ data: [sampleUser] }) },
-        { name: 'adminListFeedback', impl: async () => ({ data: [] }) },
-      ]);
-      await TestBed.configureTestingModule({
-        imports: [AdminPageComponent, MatDialogModule],
-        providers: [{ provide: Functions, useValue: {} }, provideRouter([])],
-      }).compileComponents();
-      fixture = TestBed.createComponent(AdminPageComponent);
-      component = fixture.componentInstance;
-      dialogOpenSpy = vi.spyOn(
-        fixture.debugElement.injector.get(MatDialog),
-        'open'
-      );
+    it('should open the details dialog when a non-actions cell of a rendered row is clicked', async () => {
+      // given — render a row so the template (click) wiring is exercised
+      await createComponent([sampleUser]);
       dialogOpenSpy.mockReturnValue(stubDialogRef(undefined));
-      fixture.detectChanges();
-      await fixture.whenStable();
-      fixture.detectChanges();
 
-      // Pick a non-actions cell so we don't hit the stopPropagation handler.
+      // when — pick a non-actions cell so we don't hit the stopPropagation handler
       const nameCell = fixture.nativeElement.querySelector(
         'mat-row.clickable-row mat-cell.mat-column-displayName'
       ) as HTMLElement | null;
       expect(nameCell).toBeTruthy();
       if (!nameCell) return;
-
       nameCell.dispatchEvent(
         new MouseEvent('click', { bubbles: true, cancelable: true })
       );
 
+      // then
       expect(dialogOpenSpy).toHaveBeenCalledWith(
         UserDetailsDialogComponent,
         expect.objectContaining({ data: sampleUser })
       );
     });
 
-    it('opens only the delete dialog (not the details dialog) when the row delete button is clicked', async () => {
-      // Render a single user so a row + delete button exist in the table.
-      setupCallables([
-        { name: 'adminListUsers', impl: async () => ({ data: [sampleUser] }) },
-        { name: 'adminListFeedback', impl: async () => ({ data: [] }) },
-      ]);
-      await TestBed.configureTestingModule({
-        imports: [AdminPageComponent, MatDialogModule],
-        providers: [{ provide: Functions, useValue: {} }, provideRouter([])],
-      }).compileComponents();
-      fixture = TestBed.createComponent(AdminPageComponent);
-      component = fixture.componentInstance;
-      dialogOpenSpy = vi.spyOn(
-        fixture.debugElement.injector.get(MatDialog),
-        'open'
-      );
-      // Stub the delete-confirmation dialog so the (await) delete flow does
-      // not actually call the (mocked-missing) adminDeleteUser callable.
+    it('should open only the delete dialog (not the details dialog) when the row delete button is clicked', async () => {
+      // given — render a single user so a row + delete button exist
+      await createComponent([sampleUser]);
+      // Stub the delete-confirmation dialog so the (await) delete flow does not
+      // call the (mocked-missing) adminDeleteUser callable.
       dialogOpenSpy.mockReturnValue(stubDialogRef(undefined));
-      fixture.detectChanges();
-      await fixture.whenStable();
-      fixture.detectChanges();
 
+      // when
       const deleteButton = fixture.nativeElement.querySelector(
         'mat-cell.mat-column-actions button'
       ) as HTMLButtonElement | null;
       expect(deleteButton).toBeTruthy();
       if (!deleteButton) return;
-
       deleteButton.dispatchEvent(
         new MouseEvent('click', { bubbles: true, cancelable: true })
       );
       await fixture.whenStable();
 
-      // The delete dialog should open ...
+      // then — the delete dialog opens ...
       expect(dialogOpenSpy).toHaveBeenCalledWith(
         DeleteUserDialogComponent,
         expect.objectContaining({ data: sampleUser })
       );
-      // ... but the row click handler must NOT fire (stopPropagation works).
+      // ... but the row click handler must NOT fire (stopPropagation works)
       expect(dialogOpenSpy).not.toHaveBeenCalledWith(
         UserDetailsDialogComponent,
         expect.anything()
